AWS/Lambda/CfLogs_to_ElasticSearch/requirements.txt
aws_xray_sdk >= 0.92.2
python_geoip >= 1.2
python_geoip_geolite2
requests >= 2.13.0
requests_aws4auth >= 0.9
ua_parser >= 0.7.3
